+++ parser_library/src/context/hlasm_context.h
@@ -20,6 +20,7 @@
 #include <map>
 #include <memory>
 #include <set>
+#include <string>
 #include <unordered_set>
 #include <vector>
 
@@ -114,6 +115,8 @@ class hlasm_context
     const utils::resource::resource_location* shared_resource_location(utils::resource::resource_location&&);
 
     processing_frame_tree m_stack_tree;
+
+    std::string m_title_name;
 
 public:
     hlasm_context(utils::resource::resource_location file_loc = utils::resource::resource_location(""),
@@ -331,6 +334,9 @@ public:
     const opcode_t* find_opcode_mnemo(id_index name, opcode_generation gen) const;
 
     opcode_generation current_opcode_generation() const { return m_current_opcode_generation; }
+
+    const std::string& get_title_name() const { return m_title_name; }
+    void set_title_name(std::string name) { m_title_name = std::move(name); }
 };
 
 bool test_symbol_for_read(const var_sym_ptr& var,